refactor confirm/cancel withdraw request backend
add Authorization and checks

#193
